requirements/default.txt
-c default_constraints.txt
eth-brownie==1.18.1
autopep8==1.6.0
pylint==2.13.5
